app/src/lambda/requirements.txt
boto3==1.26.105
amazon-textract-response-parser
requests
marshmallow
Pillow
pypdf